+++ Konsole/Menus/Menu.cs
@@ -13,74 +13,6 @@ namespace Konsole.Menus
 {
     // throw at any time to exit the menu.
 
-    public class WindowTheme
-    {
-        public ConsoleColor Background { get; set; } = ConsoleColor.DarkBlue;
-        public ConsoleColor Border { get; set; } = ConsoleColor.DarkBlue;
-        public ConsoleColor Foreground { get; set; } = ConsoleColor.Gray;
-        public bool ShowBorder { get; set; }
-        public LineThickNess BorderThickness { get; set; }
-
-        public static WindowTheme DarkBlue => new WindowTheme
-        {
-            Background = ConsoleColor.DarkBlue,
-            Border = ConsoleColor.Gray,
-            Foreground = ConsoleColor.Gray,
-            ShowBorder = true,
-            BorderThickness = LineThickNess.Single
-        };
-        public static WindowTheme Gray => new WindowTheme
-        {
-            Background = ConsoleColor.Gray,
-            Border = ConsoleColor.Black,
-            Foreground = ConsoleColor.Black,
-            ShowBorder = true,
-            BorderThickness = LineThickNess.Single
-        };
-        public static WindowTheme Black => new WindowTheme
-        {
-            Background = ConsoleColor.Black,
-            Border = ConsoleColor.Gray,
-            Foreground = ConsoleColor.Gray,
-            ShowBorder = true,
-            BorderThickness = LineThickNess.Single
-        };
-
-    }
-
-    public class Theme
-    {
-        public ConsoleColor Background { get; set; } = ConsoleColor.DarkBlue;
-        public ConsoleColor Border { get; set; } = ConsoleColor.DarkBlue;
-        public ConsoleColor Foreground { get; set; } = ConsoleColor.Gray;
-        public ConsoleColor ShortcutKeyHilite { get; set; } = ConsoleColor.White;
-        public ConsoleColor ShortcutKeyHiliteSelected { get; set; } = ConsoleColor.Red;
-        public ConsoleColor SelectedItemBackground { get; set; } = ConsoleColor.Gray;
-        public ConsoleColor SelectedItemForeground { get; set; } = ConsoleColor.DarkBlue;
-    }
-
-    public class Model
-    {
-        public IConsole Window { get; }
-        public string Title { get; }
-        public int Current { get; }
-        public int Height { get; }
-        public int Width { get; }
-        public MenuItem[] MenuItems { get; }
-        public Theme Theme { get; }
-
-        public Model(IConsole window, string title, int current, int height, int width, MenuItem[] menuItems, Theme theme)
-        {
-            Window = window;
-            Title = title;
-            Current = current;
-            Height = height;
-            Width = width;
-            MenuItems = menuItems;
-            Theme = theme;
-        }
-    }
-
     // Requirements 
     // - menu should run inline
     // - when finished, should continue below from where the menu was running.
@@ -105,9 +37,25 @@ namespace Konsole.Menus
 
         private Dictionary<ConsoleKey, int> _keyBindings = new Dictionary<ConsoleKey, int>();
 
-        // hooks for clearing screen and redrawing menu, if required
-        public Action BeforeMenu = () => { };
-        public Action AfterMenu = () => { };
+        /// <summary>
+        /// called before a menu item is called.
+        /// </summary>
+        public Action<MenuItem> BeforeMenuItem = (i) => { };
+        
+        /// <summary>
+        /// called after a menu item has completed.
+        /// </summary>
+        public Action<MenuItem> AfterMenuItem = (i) => { };
+
+        /// <summary>
+        /// Called after the menu has run, and the user has selected to exit the menu. 
+        /// </summary>
+        public Action<Menu> AfterMenu = (m) => { };
+
+        /// <summary>
+        /// Called before the menu starts running, i.e. at the start of .Run(), and before anything is rendered any of the consoles.
+        /// </summary>
+        public Action<Menu> BeforeMenu = (m) => { };
 
         /// <summary>
         /// Enable to display the shortcut key for the menu
@@ -275,10 +223,13 @@ namespace Konsole.Menus
             ConsoleState state;
             ConsoleKey cmd;
 
+            // #ADH I have a nested locker, parent Run locks as well, is this one needed then? must check...where is the unit test that proves I need this?
+            // i.e. if I remove the locker below, what is guaranteed to break and where is the test to prove that? otherwise I'm just guessing...
             lock (_locker)
             {
                 _menuConsole.CursorVisible = false;
                 state = _menuConsole.State;
+                BeforeMenu(this);
             }
             Refresh();
 
@@ -308,6 +259,7 @@ namespace Konsole.Menus
                 SetSelected(cmd);
                 RunItem(state, item);
             }
+            AfterMenu(this);
 
         }
 
@@ -324,18 +276,18 @@ namespace Konsole.Menus
             }
         }
 
-        private void RunItem(ConsoleState state, MenuItem item)
+        protected virtual void RunItem(ConsoleState state, MenuItem item)
         {
             try
             {
                 _menuConsole.State = state;
-                BeforeMenu();
+                BeforeMenuItem(item);
                 item.Action?.Invoke();
                 item.ConsoleAction?.Invoke(_output);
             }
             finally
             {
-                AfterMenu();
+                AfterMenuItem(item);
                 _menuConsole.State = state;
             }
         }
